+++ src/utils/usageData.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import reactGA from 'react-ga';
+import { ApplicationInsights } from '@microsoft/applicationinsights-web';
 import type Systeminformation from 'systeminformation';
 
 import logger from '../logging';
-import { PackageJson } from './AppTypes';
+import type { PackageJson } from './AppTypes';
 import { isDevelopment } from './environment';
 import {
     deleteIsSendingUsageData,
@@ -17,9 +17,7 @@ import {
     persistIsSendingUsageData,
 } from './persistentStore';
 
-const trackId = 'UA-22498474-5';
-const categoryName = () =>
-    isDevelopment ? `${appJson.name}-dev` : appJson.name;
+const instrumentationKey = '4b8b1a39-37c7-479e-a684-d4763c7c647c';
 
 interface EventAction {
     action: string;
@@ -27,8 +25,8 @@ interface EventAction {
 }
 
 let initialized = false;
-let appJson: PackageJson;
 let eventQueue: EventAction[] = [];
+let insights: ApplicationInsights;
 
 /**
  * Initialize instance to send user data
@@ -37,45 +35,50 @@ let eventQueue: EventAction[] = [];
  * @returns {Promise<void>} void
  */
 export const init = (packageJson: PackageJson) => {
-    appJson = packageJson;
+    const applicationName = isDevelopment
+        ? `${packageJson.name}-dev`
+        : packageJson.name;
+    const applicationVersion = packageJson.version;
 
     if (!getIsSendingUsageData()) return;
 
+    const accountId = getUsageDataClientId();
+
+    insights = new ApplicationInsights({
+        config: {
+            instrumentationKey,
+            accountId,
+        },
+    });
+
+    insights.loadAppInsights();
+    initialized = true;
+    insights.trackPageView({ name: applicationName });
+
+    // Add app name and version to every event
+    insights.addTelemetryInitializer(envelope => {
+        const trace = {
+            ...(envelope.ext?.trace ?? {}),
+            name: applicationName,
+        };
+        envelope.ext = { ...envelope.ext, trace };
+        envelope.data = {
+            ...envelope.data,
+            applicationName,
+            applicationVersion,
+        };
+    });
+
+    logger.debug(
+        `Application Insights for category ${applicationName} has initialized`
+    );
+
+    // Add 5 second delay to prevent inital rendering from beeing frozen.
     setTimeout(async () => {
-        const clientId = getUsageDataClientId();
-
-        logger.debug(`Client Id: ${clientId}`);
-
-        reactGA.initialize(trackId, {
-            debug: false,
-            titleCase: false,
-            gaOptions: {
-                storage: 'none',
-                storeGac: false,
-                clientId,
-            },
-        });
-
-        reactGA.set({
-            checkProtocolTask: null,
-            // According to Nordic Personal Data Processing Protocol for nRF Connect for Desktop
-            // https://projecttools.nordicsemi.no/confluence/display/ISMS/nRF+Connect+for+Desktop
-            // we set ip as anonymized in Google Analytics as described on page
-            // https://developers.google.com/analytics/devguides/collection/analyticsjs/field-reference#anonymizeIp
-            anonymizeIp: true,
-        });
-
-        reactGA.pageview(appJson.name);
-
         // eslint-disable-next-line global-require
         const si = require('systeminformation') as typeof Systeminformation;
         sendUsageData('architecture', (await si.osInfo()).arch);
-
-        initialized = true;
-        logger.debug(
-            `Google Analytics for category ${categoryName()} has initialized`
-        );
-    }, 5000); // Add 5 second delay to prevent inital rendering from beeing frozen.
+    }, 5_000);
 };
 
 /**
@@ -140,17 +143,18 @@ export const reset = () => {
  */
 const sendEvent = ({ action, label }: EventAction) => {
     const isSendingUsageData = getIsSendingUsageData();
-    const category = categoryName();
 
     if (isSendingUsageData) {
-        const data = JSON.stringify({ category, action, label });
-        logger.debug(`Sending usage data ${data}`);
-        reactGA.event({ category, action, label });
+        logger.debug(`Sending usage data ${action} ${label}`);
+        insights.trackEvent({
+            name: action,
+            properties: label ? { label } : undefined,
+        });
     }
 };
 
 /**
- * Send usage data event to Google Analytics
+ * Send usage data event to Application Insights
  * @param {string} action The event action
  * @param {string} label The event label
  * @returns {void}
@@ -165,15 +169,18 @@ export const sendUsageData = <T extends string>(action: T, label?: string) => {
 };
 
 /**
- * Send error usage data event to Google Analytics and also show it in the logger view
+ * Send error usage data event to Application Insights and also show it in the logger view
  * @param {string} error The event action
  * @returns {void}
  */
 export const sendErrorReport = (error: string) => {
     logger.error(error);
+    insights.trackException({
+        exception: new Error(error),
+    });
     sendUsageData(
         'Report error',
-        `${process.platform}; ${process.arch}; v${appJson?.version}; ${error}`
+        `${process.platform}; ${process.arch}; ${error}`
     );
 };